Enable receive custom popup dialog in ColumnMenuDelegate (#8)
This PR aims to allow the `ColumnMenuDelegate`, the delegate that controls the popup when the 3 points are pressed, to receive a custom dialog. Before this PR was only possible to override the default class to pass a list of `PopupMenuEntry`, an abstract class that contains a child widget and a value to be returned, passed to `showMenu` dialog. It wasn't possible to create any other type of dialog.

Now, it's possible to override the method resposible to call the popup, `showColumnMenu`, passing a custom dialog.

Important to add it's necessary to override and in the callback of the action, pass the value on the `Navigator.pop(context, returned_value)`, this value will be received by `onSelected` method.