Add leadingIcon widget beside PlutoColumnTitle icon (#10)
This PR aims to allow to create an auxiliar widget beside PlutoColumnTitle icon, that can be passed by implementing the PlutoColumnMenuDelegate.

It can be used to add a verification if the column is filter, to remove the filter, etc.

This can be controlled by adding a conditional statement using the PlutoColumn received by the method to show only in the right column. Can controlled some type of state in the onSelected method too.